Initial cut at better interpolation of buckets

+++ pkg/quantile/main_test.go
@@ -45,7 +45,7 @@ func ParseSketch(t *testing.T, dsl string) *Sketch {
 		}
 
 		s.count += int(n)
-		s.bins = append(s.bins, bin{k: k, n: uint16(n)})
+		s.bins = append(s.bins, bin{k: k, n: uint32(n)})
 		s.Basic.InsertN(c.f64(k), uint(n))
 	})
